Created repos and appropriate	files. Created basic sketch Layout.
Created Etch-a-Sketch repository and cloned it to local system. Created files index.html, style.css javascript.js and readme.txt. html conatins two divs, a header and a container for future divs created by js function. Css contains styling elements for div container aswell as future cells. Js contains code to create an entered amount of cells. Function makeRows (to be renamed to makeCells) creates cells that are sent to the main html container and display in a 1:1 ratio.

Next we need to add hover effects and click events in order to save clicks in the sketch. Also organize and clean up code in all aspects.